+++ src/draco/compression/mesh/mesh_sequential_encoder.cc
@@ -115,7 +115,7 @@ bool MeshSequentialEncoder::CompressAndEncodeIndices() {
       last_index_value = index_value;
     }
   }
-  EncodeSymbols(indices_buffer.data(), indices_buffer.size(), 1, nullptr,
+  EncodeSymbols(indices_buffer.data(), static_cast<int>(indices_buffer.size()), 1, nullptr,
                 buffer());
   return true;
 }
